Add: ✨ Registering first block
The block.json file represents the meta data of the block.

The index.assets.php is just an assets to tell what features the block is
requires in order to work such as dependency of WP core feature(s)

The index.js is used for creating the Gutenberg block. We want to use
the React in here