Add: ✨ Alternative approach for changing the background and Text Colors by using block supports
This is much easier way to do compare to the previous git commit.
The difference is that if you need to do some sort of inline styling vs
class name, then the previous commit could be much more helpful to do.
If it's more generic approach, use this approach instead