Fix: 🐛 infinity spinner loading edge case for loading spinner when user is reloading the editor page
When the user is trying to upload an image on the editor and then reload
the website very fast, then spinner loader is displaying. This is because
the WP is trying to look for the url image and id image.

By using the useEffect, we will always ensure the first render
on the page, we will check if there is no image id and if there is image URL
then we will set the attributes of id as undefined and alt image as empty string.

This is to ensure we always guarantee on the first render, we will always check the
condition first and then set the attributes.