Add: ✨ Transformations to the Team Member Block if selected the image or gallery block
The transformation will give us the flexibility to convert the image or gallery
block into the team member block.

If we select mutliple such as 3 images, then the team members block
will convert it to the 3 colunm of the image we already have for example

PS: It's important that the transform property needs to be added on the root of the block,
which means it's the parent block, and not the child block. If it's added
on the child block, then the transform would not work at all